Added new activity

+++ src/SFA.DAS.EmployerIncentives.Functions.PaymentsProcess/CollectionPeriod.cs
@@ -2,8 +2,8 @@
 {
     public class CollectionPeriod
     {
-        public int Year { get; set; }
-        public int Month { get; set; }
+        public short Year { get; set; }
+        public short Month { get; set; }
 
         public override string ToString()
         {
